mac_requirements.txt
tensorflow-macos
tensorflow-metal
loguru
huggingface
transformers
pandas
scikit-learn
matplotlib
plotly
kaleido
nltk
gensim
sentence_transformers
datasets
spacy
wordcloud
tensorflow_datasets
optuna
mlflow
fastapi
uvicorn
python-multipart
faiss-cpu
evaluate
seqeval